add usePromptActions
