+++ apps/chat/src/components/Promptbar/components/PromptDialogs.tsx
@@ -1,0 +1,112 @@
+import { useMemo } from 'react';
+
+import { useTranslation } from 'next-i18next';
+
+import { usePromptActions } from '@/src/hooks/usePromptActions';
+
+import { getIdWithoutRootPathSegments, isRootId } from '@/src/utils/app/id';
+import { defaultMyItemsFilters } from '@/src/utils/app/search';
+
+import { FeatureType } from '@/src/types/common';
+import { Prompt } from '@/src/types/prompt';
+import { SharingType } from '@/src/types/share';
+import { Translation } from '@/src/types/translation';
+
+import { useAppSelector } from '@/src/store/hooks';
+import { PromptsSelectors } from '@/src/store/prompts/prompts.reducers';
+
+import { PublishModal } from '@/src/components/Chat/Publish/PublishWizard';
+import { ConfirmDialog } from '@/src/components/Common/ConfirmDialog';
+
+import { MoveToFolderModal } from '../../Common/MoveToFolderModal';
+
+import { PublishActions } from '@epam/ai-dial-shared';
+
+interface Props {
+  prompt: Prompt;
+  isDeleteDialog?: boolean;
+  isUnshareDialog?: boolean;
+  publishPromptAction?: PublishActions;
+  moveTo?: { isOpen: boolean; isMobileOnly: boolean };
+  onCloseModals: () => void;
+}
+
+export const PromptDialogs: React.FC<Props> = ({
+  prompt,
+  isDeleteDialog,
+  isUnshareDialog,
+  publishPromptAction,
+  moveTo,
+  onCloseModals,
+}) => {
+  const { t } = useTranslation(Translation.PromptBar);
+
+  const filteredFoldersSelector = useMemo(
+    () =>
+      PromptsSelectors.selectFilteredFolders(defaultMyItemsFilters, '', true),
+    [],
+  );
+
+  const folders = useAppSelector(filteredFoldersSelector);
+
+  const { handleMoveToFolder, handleDelete, handleUnshare } =
+    usePromptActions(prompt);
+
+  return (
+    <>
+      {moveTo && moveTo.isOpen && (
+        <div className={moveTo.isMobileOnly ? 'md:hidden' : ''}>
+          <MoveToFolderModal
+            folders={folders}
+            onMoveToFolder={handleMoveToFolder}
+            onClose={onCloseModals}
+            featureType={FeatureType.Prompt}
+          />
+        </div>
+      )}
+      {publishPromptAction && (
+        <PublishModal
+          entity={prompt}
+          type={SharingType.Prompt}
+          isOpen
+          onClose={onCloseModals}
+          publishAction={publishPromptAction}
+          defaultPath={
+            publishPromptAction === PublishActions.DELETE &&
+            !isRootId(prompt.folderId)
+              ? getIdWithoutRootPathSegments(prompt.folderId)
+              : undefined
+          }
+        />
+      )}
+      {isDeleteDialog && (
+        <ConfirmDialog
+          isOpen
+          heading={t('Confirm deleting prompt')}
+          description={`${t('Are you sure that you want to delete a prompt?')}${t(
+            prompt.isShared
+              ? '\nDeleting will stop sharing and other users will no longer see this prompt.'
+              : '',
+          )}`}
+          confirmLabel={t('Delete')}
+          cancelLabel={t('Cancel')}
+          onClose={handleDelete}
+        />
+      )}
+      {isUnshareDialog && (
+        <ConfirmDialog
+          isOpen
+          heading={t('Confirm unsharing: {{promptName}}', {
+            promptName: prompt.name,
+          })}
+          description={
+            t('Are you sure that you want to unshare this prompt?') ?? ''
+          }
+          confirmLabel={t('Unshare')}
+          cancelLabel={t('Cancel')}
+          onClose={handleUnshare}
+        />
+      )}
+    </>
+  );
+};
