+++ apps/chat/src/components/Promptbar/components/ViewPromptButtons.tsx
@@ -14,44 +14,21 @@ import { useCallback, useMemo, useState } from 'react';
 
 import { useTranslation } from 'next-i18next';
 
-import { isEntityNameOnSameLevelUnique } from '@/src/utils/app/common';
-import { getNextDefaultName } from '@/src/utils/app/folders';
-import {
-  getIdWithoutRootPathSegments,
-  getPromptRootId,
-  isMyEntity,
-} from '@/src/utils/app/id';
-import { regeneratePromptId } from '@/src/utils/app/prompts';
+import { usePromptActions } from '@/src/hooks/usePromptActions';
+
+import { isMyEntity } from '@/src/utils/app/id';
 import { isEntityIdPublic } from '@/src/utils/app/publications';
-import { defaultMyItemsFilters } from '@/src/utils/app/search';
-import { constructPath, isRootId } from '@/src/utils/app/shared-utils';
 
 import { FeatureType } from '@/src/types/common';
-import { MoveToFolderProps } from '@/src/types/folder';
 import { Prompt } from '@/src/types/prompt';
-import { SharingType } from '@/src/types/share';
 import { Translation } from '@/src/types/translation';
 
-import { ChatActions } from '@/src/store/chat/chat.reducer';
-import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
-import { ImportExportActions } from '@/src/store/import-export/importExport.reducers';
-import {
-  PromptsActions,
-  PromptsSelectors,
-} from '@/src/store/prompts/prompts.reducers';
+import { useAppSelector } from '@/src/store/hooks';
 import { SettingsSelectors } from '@/src/store/settings/settings.selectors';
-import { ShareActions } from '@/src/store/share/share.reducers';
-import { UIActions } from '@/src/store/ui/ui.reducers';
-import { UISelectors } from '@/src/store/ui/ui.selectors';
 
-import { DEFAULT_FOLDER_NAME } from '@/src/constants/default-ui-settings';
-import { PINNED_PROMPTS_SECTION_NAME } from '@/src/constants/sections';
-
-import { PublishModal } from '@/src/components/Chat/Publish/PublishWizard';
-import { MoveToFolderModal } from '@/src/components/Common/MoveToFolderModal';
 import Tooltip from '@/src/components/Common/Tooltip';
 
-import { PromptConfirmDialogs } from './PromptConfirmDialogs';
+import { PromptDialogs } from './PromptDialogs';
 
 import UnpublishIcon from '@/public/images/icons/unpublish.svg';
 import { PublishActions } from '@epam/ai-dial-shared';
@@ -90,29 +67,15 @@ interface Props {
 }
 
 export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
-  const { t } = useTranslation(Translation.PromptBar);
-
-  const dispatch = useAppDispatch();
-
-  const [isShowMoveToModal, setIsShowMoveToModal] = useState(false);
+  const [isMoveTo, setIsMoveTo] = useState(false);
   const [publishPromptAction, setPublishPromptAction] =
     useState<PublishActions>();
   const [isDeleting, setIsDeleting] = useState(false);
   const [isUnsharing, setIsUnsharing] = useState(false);
 
-  const filteredFoldersSelector = useMemo(
-    () =>
-      PromptsSelectors.selectFilteredFolders(defaultMyItemsFilters, '', true),
-    [],
-  );
-  const collapsedSectionsSelector = useMemo(
-    () => UISelectors.selectCollapsedSections(FeatureType.Prompt),
-    [],
-  );
+  const { handleDuplicate, handleExport, handleInfo, handleShare } =
+    usePromptActions(prompt);
 
-  const folders = useAppSelector(filteredFoldersSelector);
-  const allPrompts = useAppSelector(PromptsSelectors.selectPrompts);
-  const collapsedSections = useAppSelector(collapsedSectionsSelector);
   const isPublishingEnabled = useAppSelector((state) =>
     SettingsSelectors.selectIsPublishingEnabled(state, FeatureType.Prompt),
   );
@@ -120,83 +83,11 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
     SettingsSelectors.isSharingEnabled(state, FeatureType.Prompt),
   );
 
-  const handleClosePublishModal = useCallback(() => {
-    setPublishPromptAction(undefined);
-  }, []);
-
-  const handleMoveToFolder = useCallback(
-    ({ folderId, isNewFolder }: MoveToFolderProps) => {
-      const promptRootId = getPromptRootId();
-      const folderPath = (
-        isNewFolder
-          ? getNextDefaultName(
-              t(DEFAULT_FOLDER_NAME),
-              folders.filter((f) => f.folderId === promptRootId),
-            )
-          : folderId
-      ) as string;
-
-      if (
-        !isEntityNameOnSameLevelUnique(
-          prompt.name,
-          { ...prompt, folderId: folderPath },
-          allPrompts,
-        )
-      ) {
-        dispatch(
-          UIActions.showErrorToast(
-            t('Prompt with name "{{name}}" already exists in this folder.', {
-              ns: Translation.PromptBar,
-              name: prompt.name,
-            }),
-          ),
-        );
-
-        return;
-      }
-
-      if (isNewFolder) {
-        dispatch(
-          PromptsActions.createFolder({
-            name: folderPath,
-            parentId: getPromptRootId(),
-          }),
-        );
-      }
-
-      dispatch(
-        UIActions.setCollapsedSections({
-          featureType: FeatureType.Prompt,
-          collapsedSections: collapsedSections.filter(
-            (section) => section !== PINNED_PROMPTS_SECTION_NAME,
-          ),
-        }),
-      );
-      const newFolderId = isNewFolder
-        ? constructPath(getPromptRootId(), folderPath)
-        : folderPath;
-
-      dispatch(
-        PromptsActions.updatePrompt({
-          id: prompt.id,
-          values: {
-            folderId: newFolderId,
-          },
-        }),
-      );
-      dispatch(
-        PromptsActions.setSelectedPrompt({
-          promptId: regeneratePromptId({ ...prompt, folderId: newFolderId }).id,
-        }),
-      );
-      dispatch(PromptsActions.uploadPromptSuccess({ prompt: null }));
-    },
-    [allPrompts, collapsedSections, dispatch, folders, prompt, t],
-  );
-
-  const handleCloseConfirmDialogs = useCallback(() => {
+  const handleCloseDialogs = useCallback(() => {
     setIsDeleting(false);
     setIsUnsharing(false);
+    setPublishPromptAction(undefined);
+    setIsMoveTo(false);
   }, []);
 
   const isPublic = isEntityIdPublic(prompt);
@@ -216,22 +107,14 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
         display: true,
         dataQa: 'export-prompt',
         Icon: IconFileArrowRight,
-        onClick: () => {
-          dispatch(
-            ImportExportActions.exportPrompt({
-              id: prompt.id,
-            }),
-          );
-        },
+        onClick: handleExport,
       },
       {
         tooltip: 'Duplicate prompt',
         display: true,
         dataQa: 'duplicate-prompt',
         Icon: IconCopy,
-        onClick: () => {
-          dispatch(PromptsActions.duplicatePrompt(prompt));
-        },
+        onClick: handleDuplicate,
       },
       {
         tooltip: 'Move prompt',
@@ -239,7 +122,7 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
         dataQa: 'move-prompt',
         Icon: IconFolderShare,
         onClick: () => {
-          setIsShowMoveToModal(true);
+          setIsMoveTo(true);
         },
       },
       {
@@ -247,9 +130,7 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
         display: isMyPrompt && isSharingEnabled,
         dataQa: 'share-prompt',
         Icon: IconUserShare,
-        onClick: () => {
-          setIsUnsharing(true);
-        },
+        onClick: handleShare,
       },
       {
         tooltip: 'Unshare prompt',
@@ -257,12 +138,7 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
         dataQa: 'unshare-prompt',
         Icon: IconUserX,
         onClick: () => {
-          dispatch(
-            ShareActions.share({
-              featureType: FeatureType.Prompt,
-              resourceId: prompt.id,
-            }),
-          );
+          setIsUnsharing(true);
         },
       },
       {
@@ -299,20 +175,30 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
         display: true,
         dataQa: 'info-prompt',
         Icon: IconInfoCircle,
-        onClick: () => {
-          dispatch(ChatActions.getEntityInfo({ entityInfo: prompt }));
-        },
+        onClick: handleInfo,
       },
     ],
     [
-      dispatch,
+      handleDuplicate,
+      handleExport,
+      handleInfo,
+      handleShare,
       isMyPrompt,
       isPublic,
       isPublishingEnabled,
       isSharingEnabled,
       onEditMode,
-      prompt,
+      prompt.isShared,
+      prompt.sharedWithMe,
     ],
+  );
+
+  const moveToModel = useMemo(
+    () => ({
+      isOpen: isMoveTo,
+      isMobileOnly: false,
+    }),
+    [isMoveTo],
   );
 
   return (
@@ -320,36 +206,13 @@ export const ViewPromptButtons: React.FC<Props> = ({ prompt, onEditMode }) => {
       {promptItems.map(({ display, ...props }) =>
         display ? <PromptIconBtn key={props.tooltip} {...props} /> : null,
       )}
-      {isShowMoveToModal && (
-        <MoveToFolderModal
-          folders={folders}
-          onMoveToFolder={handleMoveToFolder}
-          onClose={() => {
-            setIsShowMoveToModal(false);
-          }}
-          featureType={FeatureType.Prompt}
-        />
-      )}
-      {publishPromptAction && (
-        <PublishModal
-          entity={prompt}
-          type={SharingType.Prompt}
-          isOpen
-          onClose={handleClosePublishModal}
-          publishAction={publishPromptAction}
-          defaultPath={
-            publishPromptAction === PublishActions.DELETE &&
-            !isRootId(prompt.folderId)
-              ? getIdWithoutRootPathSegments(prompt.folderId)
-              : undefined
-          }
-        />
-      )}
-      <PromptConfirmDialogs
+      <PromptDialogs
         prompt={prompt}
         isDeleteDialog={isDeleting}
         isUnshareDialog={isUnsharing}
-        onResolve={handleCloseConfirmDialogs}
+        publishPromptAction={publishPromptAction}
+        onCloseModals={handleCloseDialogs}
+        moveTo={moveToModel}
       />
     </>
   );
